refactor: move i18n functionality from tools to i18n
