+++ package.json
@@ -82,6 +82,7 @@
     "packages/icons",
     "packages/icons-tnt",
     "packages/ie11",
+    "packages/i18n",
     "packages/tools",
     "packages/playground"
   ]
