+++ packages/i18n/.npmignore
@@ -1,0 +1,14 @@
+*.css
+dist/resources
+dist/test-resources
+node_modules/
+src/
+test/
+bundle.*.js
+.eslintrc.js
+.eslintignore
+main.*js
+rollup*
+serve.json
+wdio.conf.js
+!core
