+++ packages/i18n/.npmrc
@@ -1,0 +1,2 @@
+# Enforce public npm registry
+registry = https://registry.npmjs.org/
