+++ packages/i18n/languages-meta.js
@@ -1,0 +1,55 @@
+const languagesMeta = {
+	"default": "en",
+	"all": [
+		"ar",
+		"bg",
+		"ca",
+		"cs",
+		"cy",
+		"da",
+		"de",
+		"el",
+		"en",
+		"en_GB",
+		"en_US_sappsd",
+		"en_US_saprigi",
+		"en_US_saptrc",
+		"es",
+		"es_MX",
+		"et",
+		"fi",
+		"fr",
+		"fr_CA",
+		"hi",
+		"hr",
+		"hu",
+		"in",
+		"it",
+		"iw",
+		"ja",
+		"kk",
+		"ko",
+		"lt",
+		"lv",
+		"ms",
+		"nl",
+		"no",
+		"pl",
+		"pt_PT",
+		"pt",
+		"ro",
+		"ru",
+		"sh",
+		"sk",
+		"sl",
+		"sv",
+		"th",
+		"tr",
+		"uk",
+		"vi",
+		"zh_CN",
+		"zh_TW",
+	],
+};
+
+module.exports = languagesMeta;
