+++ packages/tools/package.json
@@ -30,6 +30,7 @@
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^11.2.0",
     "@rollup/plugin-url": "^6.0.0",
+    "@ui5/webcomponents-i18n": "1.0.0-rc.12",
     "@wdio/cli": "^7.2.0",
     "@wdio/dot-reporter": "^7.2.0",
     "@wdio/local-runner": "^7.2.0",
@@ -66,7 +67,6 @@
     "postcss-cli": "^8.3.1",
     "postcss-combine-duplicated-selectors": "^10.0.2",
     "postcss-import": "^14.0.0",
-    "properties-reader": "^2.2.0",
     "puppeteer": "^8.0.0",
     "recursive-readdir": "^2.2.2",
     "rimraf": "^3.0.2",
